# PERF: Improve performance with copy=True and different dtype