- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

This cuts runtime roughly in half when dtype is different than the dtype from values

```
# main
%timeit DataFrame(arr, dtype="int32", copy=True)
6.01 ms ± 37.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# pr
%timeit DataFrame(arr, dtype="int32", copy=True)
3.54 ms ± 25.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```